# swagger.yaml
swagger: "2.0"
info:
  description: "1Fox Content Service"
  version: "1.0.0"
  title: "1Fox Content Service"
  contact:
    email: "dev.1foxglobal@gmail.com"
host: localhost:8084
basePath: "/v1"
...
schemes:
  - "http"
paths:
  /users/login:
    post:
      tags:
        - "Access Service"
      summary: "Add a new user"
      description: ""
      operationId: "addUser"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Schema to add a user"
          required: true
          schema:
            $ref: "#/definitions/UserResponse"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "object"
            properties:
              accessToken:
                type: "string"
        "400":
          description: "Validation failed"
  /users:
    get:
      description: "List all users"
      tags:
        - "Access Service"
      consumes:
        - "application/json"
      produces":
        - "application/json"
      summary: "List users API"
      responses:
        "200":
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/UserResponse"
        "422":
          description: "Unprocessable Entity"
          schema:
            $ref: "#/definitions/errs.ErrorResponse"
        "500":
          "description": "Internal Server Error"
          "schema":
            "$ref": "#/definitions/errs.ErrorResponse"
    post:
      description: "Create a user"
      consumes:
        "application/json"

      produces:
        "application/json"

      tags:
        - "Access Service"

      summary: "Create user API"
      parameters:

        description: "Create user object"
        name: "user"
        in: "body"
        required: true
        schema:
          $ref: "#/definitions/CreateUserRequest"



      responses:
        "200":
          description: "User was created successfully"
          schema:
            "$ref": "#/definitions/UserResponse"


        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/errs.ErrorResponse"


        "409":
          description: "User already exists"
          schema:
            $ref: "#/definitions/errs.ErrorResponse"


        "422":
          description: "Unprocessable Entity"
          schema:
            $ref: "#/definitions/errs.ErrorResponse"


        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/errs.ErrorResponse"
  /users/{id}:
    get:
      description: "Get user information"
      consumes:
        "application/json"
      produces:
        "application/json"
      tags:
        - "Access Service"
      summary: "Get user information API"
      parameters:

        type: "string"
        description: "User ID"
        name: "user_id"
        in: "path"
        required: true


      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/UserResponse"


        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/errs.ErrorResponse"


        "404":
          description: "Not Found"
          schema:
            $ref: "#/definitions/errs.ErrorResponse"


        "422":
          description: "Unprocessable Entity"
          schema:
            $ref: "#/definitions/errs.ErrorResponse"


        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/errs.ErrorResponse"
    put:
      description: "Update specific user information"
      consumes:
        "application/json"
      produces:
        "application/json"
      tags:
        - "Access Service"
      summary: "Update user information API"
      parameters:
        type: "string"
        description: "User ID"
        name: "user_id"
        in: "path"
        required: true
        description: "Update user object"
        name: "user"
        in: "body"
        required: true
        schema:
          $ref: "#/definitions/UpdateUserRequest"
      responses:
        "200":
          description: "User Information was updated successfully"
          schema:
            $ref: "#/definitions/UserResponse"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/errs.ErrorResponse"
        "404":
          description: "Not Found"
          schema:
            $ref: "#/definitions/errs.ErrorResponse"
        "422":
          description: "Unprocessable Entity"
          schema:
            $ref: "#/definitions/errs.ErrorResponse"
        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/errs.ErrorResponse"
    delete:
      description: "Delete user"
      consumes:
        "application/json"
      produces:
        "application/json"
      tags:
        - "Access Service"
      summary: "Delete user API"
      parameters:
        type: "string"
        description: "User ID"
        name: "user_id"
        in: "path"
        required: true
      responses:
        "204":
          description: "User was deleted successfully"
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/errs.ErrorResponse"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/errs.ErrorResponse"
        "422":
          description: "Unprocessable Entity"
          schema:
            $ref: "#/definitions/errs.ErrorResponse"
        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/errs.ErrorResponse"
...
definitions:
...
  CreateUserRequest:
    type: "object"
    required:
      - email
      - language
      - name
      - password
      - role_id
      - translation_source_languages
      - translation_target_languages
    properties:
      email:
        type: "string"
      language:
        type: "string"
      name:
        type: "string"
      password:
        type: "string"
      role_id:
        type: "string"
      translation_source_languages:
        type: "array"
        items:
          type: "string"
      translation_target_languages:
        type: "array"
        items:
          type: "string"
  RoleResponse:
    type: "object"
    properties:
      code:
        type: "string"
      id:
        type: "string"
      name:
        type: "string"
  UpdateUserRequest:
    type: "object"
    required:
      - "language"
      - "name"
    properties:
      language:
        type: "string"
      name:
        type: "string"
      role_id:
        type: "string"
      translation_source_languages:
        type: "array"
        items:
          type: "string"
      translation_target_languages:
        type: "array"
        items:
          type: "string"
  UserResponse:
    type: "object"
    properties:
      email:
        type: "string"
      id:
        type: "string"
      language:
        type: "string"
      name:
        type: "string"
      role:
        $ref: "#/definitions/RoleResponse"
      translation_source_languages:
        type: "array"
        items:
          type: "string"
      translation_target_languages:
        type: "array"
        items:
          type: "string"
  errs.ErrorResponse:
    type: "object"
    properties:
      code:
        type: "integer"
      error:
        type: "string"
      errors:
        type: "array"
        items:
          $ref: "#/definitions/validation.ErrorMsg"
      message:
        type: "string"
  validation.ErrorMsg:
    type: "object"
    properties:
      field:
        type: "string"
      message:
        type: "string"
...